Rename update_inker to update_linker
